organise directory and code, updated README
created a data directory to store the different data created,
to avoid cluttering the root directory.

split out the google api query code into a single notebook, to
avoid accidentally rerunning it when running the main code.

modified the folium center latlon and zoom for better visualization
